# spec/openapi.yaml
openapi: 3.0.3
info:
  title: Facturacion Simple Cliente SaaS API
  version: 0.1.0
  description: |
    API propia del SaaS Facturacion Simple Cliente.

    Este contrato cubre la aplicacion operativa mobile-first y el backoffice interno minimo.
    La autoridad fiscal sigue siendo `facturacion-electronica` Ventax.

    Reglas base:
    - El operador emite para un unico facturador asociado.
    - No se guardan borradores.
    - Los precios unitarios son enteros en PYG e incluyen IVA.
    - Base imponible e IVA se redondean por linea.
    - La liquidacion por tasa suma lineas ya redondeadas.
    - El link publico usa token opaco, no CDC puro ni ID secuencial.
servers:
  - url: /api/v1
...
security:
  - BearerAuth: []
paths:
...
  /facturas/preview:
    post:
      tags: [Facturas]
      summary: Validar y calcular totales antes de emitir
      description: |
        No persiste borrador. Solo devuelve validaciones, lineas normalizadas y totales.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FacturaPreviewRequest"
      responses:
        "200":
          description: Preview calculado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FacturaPreviewResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Usuario sin configuracion operativa completa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
  /facturas:
    get:
      tags: [Facturas]
      summary: Listar facturas y notas emitidas del facturador del usuario
      parameters:
        - in: query
          name: tipo
          schema:
            $ref: "#/components/schemas/DocumentoTipo"
        - in: query
          name: estado
          schema:
            $ref: "#/components/schemas/DocumentoEstado"
        - in: query
          name: desde
          schema:
            type: string
            format: date
        - in: query
          name: hasta
          schema:
            type: string
            format: date
        - $ref: "#/components/parameters/Q"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Lista de documentos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentoListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
    post:
      tags: [Facturas]
      summary: Encolar emision de Factura Electronica contado o credito
      description: |
        Valida y persiste la factura operativa sin crear borrador.
        La llamada a `facturacion-electronica` se procesa por outbox/worker usando
        `external_ref` idempotente. La respuesta inicial puede quedar `EMITIENDO`;
        el numero fiscal, CDC y estado final vienen del backend fiscal.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FacturaEmitRequest"
      responses:
        "201":
          description: Factura emitida o en proceso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentoResponse"
        "400":
          description: Header `Idempotency-Key` ausente o invalido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Idempotencia o conflicto de estado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
  /facturas/{documentoId}:
    get:
      tags: [Facturas]
      summary: Obtener detalle de factura o nota
      parameters:
        - $ref: "#/components/parameters/DocumentoId"
      responses:
        "200":
          description: Documento encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentoResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /facturas/{documentoId}/refresh-status:
    post:
      tags: [Facturas]
      summary: Refrescar estado fiscal desde Ventax FE
      parameters:
        - $ref: "#/components/parameters/DocumentoId"
      responses:
        "200":
          description: Estado actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentoResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Documento sin CDC fiscal para consultar
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Backend fiscal no disponible o rechazo la consulta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "504":
          description: Timeout al consultar backend fiscal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /facturas/{documentoId}/retry-emission:
    post:
      tags: [Facturas]
      summary: Reintentar emision fiscal recuperable
      description: |
        Reactiva el outbox de una factura en `PENDIENTE_SIFEN` o `ERROR_TEMPORAL`
        usando el mismo `external_ref` idempotente. No crea una nueva factura
        operativa ni permite reintentar estados finales.
      parameters:
        - $ref: "#/components/parameters/DocumentoId"
      responses:
        "200":
          description: Reintento encolado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentoResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Documento no recuperable o sin outbox reintentable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /facturas/{documentoId}/cancelar:
    post:
      tags: [Facturas]
      summary: Cancelar o anular documento elegible
      description: |
        Solo permite solicitar cancelacion de facturas `EMITIDA` con CDC fiscal.
        El SaaS delega la validacion fiscal final al backend `facturacion-electronica`
        mediante `/evento/cancelar` y guarda la respuesta como snapshot/auditoria.
      parameters:
        - $ref: "#/components/parameters/DocumentoId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelacionRequest"
      responses:
        "200":
          description: Cancelacion procesada y documento marcado como ANULADA
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentoResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Documento no elegible para cancelacion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
        "502":
          description: Backend fiscal rechazo la cancelacion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "504":
          description: Timeout al cancelar en backend fiscal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /facturas/{documentoId}/nota-credito:
    post:
      tags: [NotasCredito]
      summary: Emitir Nota de Credito Electronica desde factura elegible
      description: |
        Emite una NCE total en el MVP. Solo acepta facturas `EMITIDA` con CDC
        y bloquea una segunda NCE total sobre la misma factura.
      parameters:
        - $ref: "#/components/parameters/DocumentoId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotaCreditoEmitRequest"
      responses:
        "201":
          description: NCE emitida o en proceso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentoResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Factura no elegible o ya tiene NCE total
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
        "502":
          description: Backend fiscal rechazo la NCE
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "504":
          description: Timeout al emitir NCE contra backend fiscal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /backoffice/users:
    post:
      tags: [Backoffice]
      summary: Crear usuario operativo o interno
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BackofficeUserCreateRequest"
      responses:
        "201":
          description: Usuario creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackofficeUserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Permiso de backoffice requerido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username ya existe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /backoffice/users/{userId}/reset-password:
    post:
      tags: [Backoffice]
      summary: Desbloquear usuario y definir o regenerar password temporal
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BackofficePasswordResetRequest"
      responses:
        "200":
          description: Password temporal generado o actualizado; sesiones activas revocadas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackofficeUserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Permiso de backoffice requerido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /backoffice/users/{userId}/operation-config:
    put:
      tags: [Backoffice]
      summary: Asociar configuracion operativa unica al usuario
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserOperationConfigRequest"
      responses:
        "200":
          description: Configuracion actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserOperationConfig"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Permiso de backoffice requerido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Usuario o configuracion operativa no encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Q:
      in: query
      name: q
      required: false
      schema:
        type: string
    Limit:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 200
    Offset:
      in: query
      name: offset
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
...
    DocumentoId:
      in: path
      name: documentoId
      required: true
      schema:
        type: string
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 120
...
  responses:
    Unauthorized:
      description: No autenticado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Error de validacion
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          example: validation_error
        message:
          type: string
          example: Datos invalidos
        details:
          type: object
          nullable: true
          additionalProperties: true
...
    DocumentoIdentidadTipo:
      type: string
      enum: [RUC, CI, PASAPORTE, CEDULA_EXTRANJERA, NO_ESPECIFICADO]
...
    IvaTipo:
      type: string
      enum: [IVA_10, IVA_5, EXENTA]
    CondicionVenta:
      type: string
      enum: [CONTADO, CREDITO]
    DocumentoTipo:
      type: string
      enum: [FACTURA, NOTA_CREDITO]
    DocumentoEstado:
      type: string
      enum: [EMITIENDO, EMITIDA, PENDIENTE_SIFEN, RECHAZADA, ERROR_OPERATIVO, ERROR_TEMPORAL, ANULADA]
    FacturaPreviewRequest:
      type: object
      required: [condicion_venta, cliente, items]
      properties:
        condicion_venta:
          $ref: "#/components/schemas/CondicionVenta"
        cliente:
          $ref: "#/components/schemas/FacturaClienteInput"
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/FacturaItemInput"
    FacturaEmitRequest:
      allOf:
        - $ref: "#/components/schemas/FacturaPreviewRequest"
        - type: object
          properties:
            requested_at:
              type: string
              format: date-time
    FacturaClienteInput:
      type: object
      required: [documento_tipo, documento, razon_social]
      properties:
        cliente_id:
          type: string
          nullable: true
        documento_tipo:
          $ref: "#/components/schemas/DocumentoIdentidadTipo"
        documento:
          type: string
        razon_social:
          type: string
        direccion:
          type: string
          nullable: true
        telefono:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
    FacturaItemInput:
      type: object
      required: [cantidad, descripcion, precio_unitario]
      properties:
        catalogo_item_id:
          type: string
          nullable: true
        codigo:
          type: string
          nullable: true
        descripcion:
          type: string
        cantidad:
          type: integer
          minimum: 1
        precio_unitario:
          type: integer
          minimum: 0
        iva_tipo:
          $ref: "#/components/schemas/IvaTipo"
    FacturaPreviewResponse:
      type: object
      required: [items, totals]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DocumentoItemSnapshot"
        totals:
          $ref: "#/components/schemas/DocumentoTotals"
    DocumentoItemSnapshot:
      type: object
      required: [line_no, codigo, descripcion, cantidad, precio_unitario, iva_tipo, subtotal, base_imponible, iva_monto]
      properties:
        line_no:
          type: integer
        catalogo_item_id:
          type: string
          nullable: true
        codigo:
          type: string
          nullable: true
        descripcion:
          type: string
        cantidad:
          type: integer
        precio_unitario:
          type: integer
        iva_tipo:
          $ref: "#/components/schemas/IvaTipo"
        subtotal:
          type: integer
        base_imponible:
          type: integer
          description: Redondeado por linea.
        iva_monto:
          type: integer
          description: Redondeado por linea.
    DocumentoTotals:
      type: object
      required: [subtotal, total_sin_iva, iva_5, iva_10, total_iva, total]
      properties:
        subtotal:
          type: integer
        total_sin_iva:
          type: integer
        iva_5:
          type: integer
        iva_10:
          type: integer
        total_iva:
          type: integer
        total:
          type: integer
    DocumentoResponse:
      type: object
      required: [id, tipo, estado, condicion_venta, cliente, items, totals, delivery]
      properties:
        id:
          type: string
        tipo:
          $ref: "#/components/schemas/DocumentoTipo"
        estado:
          $ref: "#/components/schemas/DocumentoEstado"
        condicion_venta:
          $ref: "#/components/schemas/CondicionVenta"
        numero_fiscal:
          type: string
          nullable: true
          example: 001-001-0000123
        cdc:
          type: string
          nullable: true
        fiscal_document_id:
          type: string
          nullable: true
        external_ref:
          type: string
          nullable: true
        cliente:
          $ref: "#/components/schemas/FacturaClienteInput"
        items:
          type: array
          items:
            $ref: "#/components/schemas/DocumentoItemSnapshot"
        totals:
          $ref: "#/components/schemas/DocumentoTotals"
        fiscal_status:
          type: object
          nullable: true
          additionalProperties: true
        documento_relacionado_id:
          type: string
          nullable: true
          description: ID de la factura original cuando el documento es una Nota de Credito.
        nce_motivo:
          type: string
          nullable: true
        delivery:
          $ref: "#/components/schemas/DeliverySummary"
        created_at:
          type: string
          format: date-time
          nullable: true
    DocumentoListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DocumentoResponse"
        total:
          type: integer
    NotaCreditoEmitRequest:
      type: object
      required: [motivo]
      properties:
        motivo:
          type: string
          maxLength: 300
          description: Motivo operativo de la NCE total. Las lineas se copian de la factura original en el MVP.
    CancelacionRequest:
      type: object
      required: [motivo]
      properties:
        motivo:
          type: string
          maxLength: 150
    DeliverySummary:
      type: object
      properties:
        public_url:
          type: string
          nullable: true
        whatsapp_url:
          type: string
          nullable: true
        email_status:
          $ref: "#/components/schemas/EmailStatus"
        artifacts:
          $ref: "#/components/schemas/ArtifactsResponse"
...
    ArtifactsResponse:
      type: object
      required: [kude_pdf, xml]
      properties:
        kude_pdf:
          $ref: "#/components/schemas/ArtifactLink"
        xml:
          $ref: "#/components/schemas/ArtifactLink"
    ArtifactLink:
      type: object
      required: [available]
      properties:
        available:
          type: boolean
        url:
          type: string
          format: uri
          nullable: true
    EmailStatus:
      type: string
      enum: [NOT_APPLICABLE, DELEGATED, SENT, FAILED, UNKNOWN]
...
    BackofficeUserCreateRequest:
      type: object
      required: [username, display_name, role]
      properties:
        username:
          type: string
        display_name:
          type: string
        role:
          type: string
          enum: [OPERADOR_FACTURACION, SOPORTE_INTERNO, ADMIN_INTERNO]
        temporary_password:
          type: string
          format: password
          nullable: true
    BackofficePasswordResetRequest:
      type: object
      properties:
        temporary_password:
          type: string
          format: password
          nullable: true
    BackofficeUserResponse:
      type: object
      required: [id, username, role, active]
      properties:
        id:
          type: string
        username:
          type: string
        display_name:
          type: string
        role:
          type: string
        active:
          type: boolean
        temporary_password:
          type: string
          format: password
          nullable: true
          description: Solo se devuelve en creacion o reset; no se persiste en claro.
    UserOperationConfigRequest:
      type: object
      required: [tenant_id, facturador_id, emisor_id, establecimiento, punto_expedicion, perfil_emision_codigo, actividad_economica_codigo]
      properties:
        tenant_id:
          type: string
          format: uuid
        facturador_id:
          type: string
          format: uuid
        emisor_id:
          type: string
        establecimiento:
          type: string
        punto_expedicion:
          type: string
        perfil_emision_codigo:
          type: string
        actividad_economica_codigo:
          type: string
    UserOperationConfig:
      allOf:
        - $ref: "#/components/schemas/UserOperationConfigRequest"
        - type: object
          required: [user_id, active]
          properties:
            user_id:
              type: string
            active:
              type: boolean